AI/app/fontmaker/requirements.txt
certifi==2020.12.5
colorama==0.4.4
cycler==0.11.0
dataclasses==0.8
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.4.0
pyparsing==3.0.4
python-dateutil==2.8.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
torch==1.10.0
tqdm==4.62.3
typing-extensions==3.10.0.2
wincertstore==0.2
opencv-python==4.6.0.66
requests==2.27.1